+++ changelog.md
@@ -24,7 +24,7 @@ Codion Change Log
 - EntityTablePanel, unknown_dependent_records i18n message moved to EntityEditPanel.
 - EntityPanel.DetailLayout.layout() now throws exception in case the panel has already been laid out.
 - EntityApplicationPanel.ApplicationLayout.layout() now throws exception in case the panel has already been laid out.
-- EntityApplicationPanel.ApplicationLayout.applicationTabPane() removed.
+- EntityApplicationPanel.ApplicationLayout.applicationTabPane() renamed tabbedPane().
 
 ## 0.17.33
 ### is.codion.common.core